project_template/project_name/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-cleanup==5.2.0
django-debug-toolbar==3.2.2
django-livereload-server==0.3.2
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
urwid==2.1.2
pydantic==1.8.2
django-ninja==0.16.1
typing_extensions==4.0.1
orjson==3.6.5
